Rename: Exception 클래스 이름 변경
Exception은 이미 있는 Exception과 헷갈린다고 해서 ExceptionName으로 변경
