--- src/main/java/exception/Exception.java
@@ -1,6 +1,0 @@
-package exception;
-
-public class Exception {
-    public static final String ALREADY_EXIST_USER = "이미 존재하는 회원입니다.";
-    public static final String INVALID_URI = "uri가 적절하지 않습니다.";
-}
