+++ src/main/java/exception/ExceptionName.java
@@ -1,0 +1,7 @@
+package exception;
+
+public final class ExceptionName {
+    public static final String ALREADY_EXIST_USER = "이미 존재하는 회원입니다.";
+    public static final String NOT_ENOUGH_USER_INFORMATION = "회원 가입을 완료하려면 모든 정보를 입력해주세요.";
+    public static final String INVALID_URI = "URI가 적절하지 않습니다.";
+}
