+++ src/test/java/service/UserServiceTest.java
@@ -10,7 +10,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static db.Database.clear;
-import static exception.Exception.ALREADY_EXIST_USER;
+import static exception.ExceptionName.ALREADY_EXIST_USER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
